api/requirements.txt
scikit-learn
sqlalchemy
pandas
tensorflow-cpu
matplotlib
mlflow
joblib
numpy==1.26.4
request